requirements.txt
python-telegram-bot[webhooks]>=20.7,<21.0
PyGithub>=2.1.1,<3.0
psycopg2-binary>=2.9.9
cryptography>=41.0.7
python-dotenv>=1.0.0
aiohttp>=3.9.1
humanize>=4.9.0
